Add a beta assembler and a binary loader
This is big news, programs can now be written in assembly!
  (see multiplication.laval)

- The assembler is used to compile assembly to binary
- The binary loader is used to load binary into a functional CPU instance

- Add some corresponding tests
- Cleanup the old non-functional python assembler
- Only works on GCC for now
- Modify main to accept an assembly program as parameter

+++ lib/impl/core_array.cpp
@@ -3,7 +3,7 @@
 #include "direction.h"
 
 
-CoreArray::CoreArray(const std::vector<size_t>& dimensions, const MemoryInterface& mem)
+CoreArray::CoreArray(const std::vector<uint16_t>& dimensions, const MemoryInterface& mem)
 {
     auto size = std::accumulate(dimensions.begin(), dimensions.end(), 1ull, std::multiplies<>());
     assert(size <= Tools::umaxof<size_t>() >> 1);
